+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Add an option `--expect-no-differences` that will generate a non-zero exit code when differences are
+  detected otherwise the tool has been updated to return `0` on success.
+
 ### [v0.02](https://github.com/realityforge/revapi-diff/tree/v0.02) (2019-02-09)
 [Full Changelog](https://github.com/realityforge/revapi-diff/compare/v0.01...v0.02)